[SD] Move some common code to utility
